requirements.txt
flask
flask-cors
numpy
pillow
scikit-learn
